+++ completion/_tex-fmt.ps1
@@ -37,6 +37,7 @@ Register-ArgumentCompleter -Native -CommandName 'tex-fmt' -ScriptBlock {
             [CompletionResult]::new('--nowrap', '--nowrap', [CompletionResultType]::ParameterName, 'Do not wrap long lines')
             [CompletionResult]::new('-j', '-j', [CompletionResultType]::ParameterName, 'Join short lines within paragraphs before wrapping')
             [CompletionResult]::new('--join', '--join', [CompletionResultType]::ParameterName, 'Join short lines within paragraphs before wrapping')
+            [CompletionResult]::new('--format-options', '--format-options', [CompletionResultType]::ParameterName, 'Put each item in multiline optional arguments on its own line')
             [CompletionResult]::new('--usetabs', '--usetabs', [CompletionResultType]::ParameterName, 'Use tabs instead of spaces for indentation')
             [CompletionResult]::new('-s', '-s', [CompletionResultType]::ParameterName, 'Process stdin as a single file, output to stdout')
             [CompletionResult]::new('--stdin', '--stdin', [CompletionResultType]::ParameterName, 'Process stdin as a single file, output to stdout')